Доработки задания 8.10
