+++ js/main.js
@@ -1,3 +1,3 @@
 import './show-images.js';
 import './show-full-image.js';
-import './upload-image.js';
+import './upload-image-form.js';
